+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 Project versioning adheres to [Semantic Versioning](http://semver.org/).
 Change log format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [9.0.0](https://github.com/locize/locize-cli/compare/v8.7.1...v9.0.0) - 2025-04-03
+
+- the format `yaml-rails-ns`/`yml-rails-ns` is the old `yaml-rails`/`yml-rails`
+- `yaml-rails`/`yml-rails` does not scope the yaml content with the namespace anymore => use `yaml-rails-ns`/`yml-rails-ns` instead
+
 ## [8.7.1](https://github.com/locize/locize-cli/compare/v8.7.0...v8.7.1) - 2025-04-02
 
 - make generated files POSIX compliant, addresses [107](https://github.com/locize/locize-cli/issues/107)